+++ packages/trellis_site/lib/trellis_site.dart
@@ -7,6 +7,7 @@ library;
 export 'src/code_highlighter.dart';
 export 'src/content_discovery.dart';
 export 'src/feed_generator.dart';
+export 'src/front_matter_date.dart';
 export 'src/front_matter_parser.dart';
 export 'src/markdown_renderer.dart';
 export 'src/navigation_builder.dart';